# openapi.yaml
openapi: 3.0.3
info:
  title: Reconciliation API
  version: RECONCILIATION_VERSION
paths:
...
  /policies:
    post:
      summary: Create a policy
      tags:
        - reconciliation.v1
      operationId: createPolicy
      description: Create a policy
      requestBody:
        $ref: "#/components/requestBodies/Policy"
      responses:
        "201":
          $ref: "#/components/responses/Policy"
        default:
          $ref: "#/components/responses/ErrorResponse"
      security:
        - Authorization:
            - reconciliation:write
    get:
      summary: List policies
      operationId: listPolicies
      tags:
        - reconciliation.v1
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QueryBuilder"
      responses:
        "200":
          $ref: "#/components/responses/Policies"
        default:
          $ref: "#/components/responses/ErrorResponse"
      security:
        - Authorization:
            - reconciliation:read
  /policies/{policyID}:
    delete:
      summary: Delete a policy
      operationId: deletePolicy
      tags:
        - reconciliation.v1
      description: Delete a policy by its id.
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "204":
          $ref: "#/components/responses/NoContent"
        default:
          $ref: "#/components/responses/ErrorResponse"
      security:
        - Authorization:
            - reconciliation:write
    get:
      summary: Get a policy
      tags:
        - reconciliation.v1
      operationId: getPolicy
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "200":
          $ref: "#/components/responses/Policy"
        default:
          $ref: "#/components/responses/ErrorResponse"
      security:
        - Authorization:
            - reconciliation:read
    patch:
      summary: Update a policy
      tags:
        - reconciliation.v1
      operationId: updatePolicy
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      requestBody:
        $ref: "#/components/requestBodies/Policy"
      responses:
        "200":
          $ref: "#/components/responses/Policy"
        default:
          $ref: "#/components/responses/ErrorResponse"
      security:
        - Authorization:
            - reconciliation:write
...
  /policies/{policyID}/reports/latest:
    get:
      summary: Get latest policy report
      tags:
        - reconciliation.v1
      operationId: getLatestPolicyReport
      parameters:
        - $ref: "#/components/parameters/PolicyID"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Report"
        default:
          $ref: "#/components/responses/ErrorResponse"
      security:
        - Authorization:
            - reconciliation:read
...
components:
  parameters:
    PageSize:
      name: pageSize
      in: query
      description: The maximum number of results to return per page.
      example: 100
      schema:
        type: integer
        format: int64
        minimum: 1
        maximum: 1000
    Cursor:
      name: cursor
      in: query
      description: |
        Parameter used in pagination requests.
        Set to the value of next for the next page of results.
        Set to the value of previous for the previous page of results.
      schema:
        type: string
        example: aHR0cHM6Ly9nLnBhZ2UvTmVrby1SYW1lbj9zaGFyZQ==
    PolicyID:
      name: policyID
      in: path
      schema:
        type: string
        format: uuid
      description: The policy ID.
      required: true
...
  responses:
    NoContent:
      description: No content
...
    Policies:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PoliciesCursorResponse"
    Policy:
      description: OK
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PolicyResponse"
...
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  requestBodies:
    Policy:
      required: true
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/PolicyRequest"
...
  schemas:
    QueryBuilder:
      type: object
      additionalProperties: true
...
    ErrorResponse:
      type: object
      required:
        - errorCode
        - errorMessage
      properties:
        errorCode:
          type: string
          example: VALIDATION
        errorMessage:
          type: string
        details:
          type: string

    # V1 Schemas
    PolicyRequest:
      type: object
      required:
        - name
        - ledgerName
        - ledgerQuery
        - paymentsPoolID
      properties:
        name:
          type: string
        ledgerName:
          type: string
        ledgerQuery:
          type: object
          additionalProperties: true
        paymentsPoolID:
          type: string
    PolicyResponse:
      type: object
      required:
        - data
      properties:
        data:
          $ref: "#/components/schemas/Policy"
    Policy:
      type: object
      required:
        - id
        - name
        - createdAt
        - ledgerName
        - ledgerQuery
        - paymentsPoolID
      properties:
        id:
          type: string
        name:
          type: string
        createdAt:
          type: string
          format: date-time
        ledgerName:
          type: string
        ledgerQuery:
          type: object
          additionalProperties: true
        paymentsPoolID:
          type: string
    PoliciesCursorResponse:
      type: object
      required:
        - cursor
      properties:
        cursor:
          $ref: "#/components/schemas/CursorOf_Policy"
    CursorOf_Policy:
      type: object
      required:
        - pageSize
        - hasMore
        - data
      properties:
        pageSize:
          type: integer
          format: int64
        hasMore:
          type: boolean
        previous:
          type: string
        next:
          type: string
        data:
          type: array
          items:
            $ref: "#/components/schemas/Policy"
...
    Report:
      type: object
      required:
        - id
        - policyID
        - periodStart
        - periodEnd
        - totalTransactions
        - matchedCount
        - matchRate
        - generatedAt
      properties:
        id:
          type: string
          format: uuid
        policyID:
          type: string
          format: uuid
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        totalTransactions:
          type: integer
          format: int64
        matchedCount:
          type: integer
          format: int64
        matchRate:
          type: number
          format: double
        anomaliesByType:
          type: object
          additionalProperties:
            type: integer
            format: int64
        generatedAt:
          type: string
          format: date-time

    # V2 Schemas
...